requirements-observability.txt
# Dependencies needed for Langfuse OTLP tracing in the Pipecat bots
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-http
